// src/gen/xai/api/v1/deferred_pb.ts
// @generated by protoc-gen-es v2.13.0 with parameter "target=ts,import_extension=.js"
// @generated from file xai/api/v1/deferred.proto (package xai_api, syntax proto3)
/* eslint-disable */

import type { GenEnum, GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { enumDesc, fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file xai/api/v1/deferred.proto.
 */
export const file_xai_api_v1_deferred: GenFile = /*@__PURE__*/
  fileDesc("Chl4YWkvYXBpL3YxL2RlZmVycmVkLnByb3RvEgd4YWlfYXBpIisKFVN0YXJ0RGVmZXJyZWRSZXNwb25zZRISCgpyZXF1ZXN0X2lkGAEgASgJIigKEkdldERlZmVycmVkUmVxdWVzdBISCgpyZXF1ZXN0X2lkGAEgASgJKl0KDkRlZmVycmVkU3RhdHVzEhsKF0lOVkFMSURfREVGRVJSRURfU1RBVFVTEAASCAoERE9ORRABEgsKB0VYUElSRUQQAhILCgdQRU5ESU5HEAMSCgoGRkFJTEVEEARCVAoLY29tLnhhaV9hcGlCDURlZmVycmVkUHJvdG9QAaICA1hYWKoCBlhhaUFwacoCBlhhaUFwaeICElhhaUFwaVxHUEJNZXRhZGF0YeoCBlhhaUFwaWIGcHJvdG8z");

/**
 * The response from the service, when creating a deferred completion request.
 *
 * @generated from message xai_api.StartDeferredResponse
 */
export type StartDeferredResponse = Message<"xai_api.StartDeferredResponse"> & {
  /**
   * The ID of this request. This ID can be used to retrieve completion results
   * later.
   *
   * @generated from field: string request_id = 1;
   */
  requestId: string;
};

/**
 * Describes the message xai_api.StartDeferredResponse.
 * Use `create(StartDeferredResponseSchema)` to create a new message.
 */
export const StartDeferredResponseSchema: GenMessage<StartDeferredResponse> = /*@__PURE__*/
  messageDesc(file_xai_api_v1_deferred, 0);

/**
 * Retrieve the deferred chat request's response with the `request_id` in
 * StartDeferredResponse.
 *
 * @generated from message xai_api.GetDeferredRequest
 */
export type GetDeferredRequest = Message<"xai_api.GetDeferredRequest"> & {
  /**
   * The ID of this request to get.
   *
   * @generated from field: string request_id = 1;
   */
  requestId: string;
};

/**
 * Describes the message xai_api.GetDeferredRequest.
 * Use `create(GetDeferredRequestSchema)` to create a new message.
 */
export const GetDeferredRequestSchema: GenMessage<GetDeferredRequest> = /*@__PURE__*/
  messageDesc(file_xai_api_v1_deferred, 1);

/**
 * Status of deferred completion request.
 *
 * @generated from enum xai_api.DeferredStatus
 */
export enum DeferredStatus {
  /**
   * Invalid status.
   *
   * @generated from enum value: INVALID_DEFERRED_STATUS = 0;
   */
  INVALID_DEFERRED_STATUS = 0,

  /**
   * The request has been processed and is available for download.
   *
   * @generated from enum value: DONE = 1;
   */
  DONE = 1,

  /**
   * The request has been processed but the content has expired and is not
   * available anymore.
   *
   * @generated from enum value: EXPIRED = 2;
   */
  EXPIRED = 2,

  /**
   * The request is still being processed.
   *
   * @generated from enum value: PENDING = 3;
   */
  PENDING = 3,

  /**
   * The request failed due to an internal service error.
   * The error message is in the `error` field of the response.
   *
   * @generated from enum value: FAILED = 4;
   */
  FAILED = 4,
}

/**
 * Describes the enum xai_api.DeferredStatus.
 */
export const DeferredStatusSchema: GenEnum<DeferredStatus> = /*@__PURE__*/
  enumDesc(file_xai_api_v1_deferred, 0);